Compromise

Documentation for Compromise. There is not much here yet. Everything is still very much a work-in-progress.

Random Doc-Strings:

Compromise.MutableMOP

Compromise.ValueArrays

Compromise.add_function!

Compromise.diff_mod!

Compromise.eval_and_diff_mod!

Compromise.eval_mod!

Compromise.lin_cons!

Compromise.scale!

Compromise.scale_eq!

Compromise.set_linear_constraints!

Compromise.unscale!

Compromise.MutableMOP — Type

MutableMOP(; num_vars, kwargs...)

Initialize a multi-objective problem with num_vars variables.

Functions

There can be exactly one (possibly vector-valued) objective function, one nonlinear equality constraint function, and one nonlinear inequality constraint function. For now, they have to be of type NonlinearFunction. You could provide these functions with the keyword-arguments objectives, nl_eq_constraints or nl_ineq_constraints or set the fields of the same name. To conveniently add user-provided functions, there are helper functions, like add_objectives!.

LinearConstraints

Box constraints are defined by the vectors lb and ub. Linear equality constraints $Ex=c$ are defined by the matrix E and the vector c. Inequality constraints read $Ax≤b$ and use A and b.

Surrogate Configuration

Use the keyword-arguments mcfg_objectives to provide an AbstractSurrogateModelConfig to define how the objectives should be modelled. By default, we assume ExactModelConfig(), which requires differentiable objectives.

Compromise.ValueArrays — Type

A struct holding values computed for or derived from an AbstractMOP.

Compromise.add_function! — Function

add_function!(func_field, mop, op, model; dim_out, backend=NoBackend())

Add the operator op to mop at func_field and use model configuration model. Keyword argument dim_out::Int is mandatory. E.g., add_function!(:objectives, mop, op, :rbf; dim_out=2) adds op as the bi-valued objective to mop.

Compromise.diff_mod! — Method

Evaluate the model gradients of mod at x and store results in mod_vals::SurrogateValueArrays.

Compromise.eval_and_diff_mod! — Method

Evaluate and differentiate mod at x and store results in mod_vals::SurrogateValueArrays.

Compromise.eval_mod! — Method

Evaluate the models mod at x and store results in mod_vals::SurrogateValueArrays.

Compromise.lin_cons! — Method

lin_cons!(residual_vector, prod_cache, constraint_data, x)

Given a linear constraint A*x .<= b or A*x .== b, compute the product A*x and store the result in prod_cache, and also compute A*x .- b and store the result in residual_vector. constraint_data should either be the tuple (A,b)::Tuple{RMat,RVec} or nothing.

Compromise.scale! — Method

Scale ξ and set x according to x = T*ξ + t.

Compromise.scale_eq! — Method

Make Aξ + b ? 0 applicable in scaled domain via A(inv(T)*x - inv(T)*t) + b ? 0.

Compromise.set_linear_constraints! — Method

set_linear_constraints!(opt, affine_vec, var_vec, mat, rhs, ctype)

Add linear (in-)equality constraints to JuMP model opt.

If ctype is :eq, then the constraints read affine_vec + mat * var_vec .== rhs.

Otherwise, they read affine_vec + mat * var_vec .<= rhs.

This helper function returns a JuMP expression for the matrix-vector-product mat*var_vec.

Compromise.unscale! — Method

Unscale x and set ξ according to ξ = inv(T)*x - inv(T)*t.