Compromise

Documentation for Compromise. There is not much here yet. Everything is still very much a work-in-progress.

Random Doc-Strings:

Compromise.compute_descent_step

Compromise.compute_normal_step

Compromise.diff_mod!

Compromise.eval_and_diff_mod!

Compromise.eval_mod!

Compromise.scale!

Compromise.scale_eq!

Compromise.set_linear_constraints!

Compromise.unscale!

Compromise.compute_descent_step — Method

compute_descent_step(
    it_index, Δ, it_stat, mop, mod, scaler, lin_cons, scaled_cons,
    vals, vals_tmp, step_vals, mod_vals, filter, step_cache, algo_opts
) :: Real

Return a criticalty value χ.

Modify step_vals.d to hold the scaled descent step at step_vals.xn.

Modify step_vals.xs to hold vals.x + step_vals.n + step_vals.d, or equivalently, step_vals.xn + step_vals.d.

Modify step_vals.fxs to hold the surrogate objective values at step_vals.xs.

The descent step is computed according to step_cache and this cache can also be modified.

Usually, the descent step is computed using the surrogate values stored in mod_vals. These should not be modified!

Compromise.compute_normal_step — Method

compute_normal_step(
    it_index, Δ, it_stat, mop, mod, scaler, lin_cons, scaled_cons,
    vals, vals_tmp, step_vals, mod_vals, filter, step_cache, algo_opts
)

Modify step_vals.n to hold the normal step at vals.x.

Modify step_vals.xn to hold vals.x + step_vals.n.

The normal step is computed according to step_cache and this cache can also be modified.

Usually, the normal step is computed using the surrogate values stored in mod_vals. These should not be modified!

Compromise.diff_mod! — Method

Evaluate the model gradients of mod at x and store results in mod_vals::SurrogateValueArrays.

Compromise.eval_and_diff_mod! — Method

Evaluate and differentiate mod at x and store results in mod_vals::SurrogateValueArrays.

Compromise.eval_mod! — Method

Evaluate the models mod at x and store results in mod_vals::SurrogateValueArrays.

Compromise.scale! — Method

Scale ξ and set x according to x = T*ξ + t.

Compromise.scale_eq! — Method

Make Aξ + b ? 0 applicable in scaled domain via A(inv(T)*x - inv(T)*t) + b ? 0.

Compromise.set_linear_constraints! — Method

Add the constraint c + A * x .?= b to opt and return a JuMP expression for A*x.

Compromise.unscale! — Method

Unscale x and set ξ according to ξ = inv(T)*x - inv(T)*t.