Compromise

Documentation for Compromise. There is not much here yet. Everything is still very much a work-in-progress.

Random Doc-Strings:

Compromise.AlgorithmOptions

Compromise.MutableMOP

Compromise.NonlinearFunctions.NonlinearParametricFunction

Compromise.ValueArrays

Compromise.add_function!

Compromise.add_nl_eq_constraints!

Compromise.add_nl_eq_constraints!

Compromise.add_nl_eq_constraints!

Compromise.add_nl_ineq_constraints!

Compromise.add_nl_ineq_constraints!

Compromise.add_nl_ineq_constraints!

Compromise.add_objectives!

Compromise.add_objectives!

Compromise.add_objectives!

Compromise.diff_mod!

Compromise.eval_and_diff_mod!

Compromise.eval_mod!

Compromise.lin_cons!

Compromise.scale!

Compromise.scale_eq!

Compromise.set_linear_constraints!

Compromise.unscale!

Compromise.NonlinearFunctions.NonlinearParametricFunction — Type

NonlinearParametricFunction(; 
    func, grads=nothing, hessians=nothing, 
    func_and_grads=nothing, func_and_grads_and_hessians=nothing,
    backend=nothing,
    func_iip=true, grads_iip=true, hessians_iip=true, 
    func_and_grads_iip=true, func_and_grads_and_hessians_iip=true)

A flexible function wrapper to conveniently query evaluations and derivatives of user provided functions.

If the user provided function is used in a derivative-free alogrithm, only func has to be provided. The flag func_iip indicates its signature: If func_iip==true, the function should mutate the target array and have signature func!(y, x, p). Otherwise, y = func(x, p).

Should gradients be needed, function handles can be provided, and the respective flags indicate the following signatures:

grads_iip == true implies grads!(Dy, x, p), otherwise Dy = grads(x, p).

hessians_iip == true implies hessians!(H, x, p), otherwise H = hessians(x, p).

func_and_grads_iip == true implies func_and_grads!(y, Dy, x, p), otherwise y, Dy = func_and_grads(x, p).

func_and_grads_and_hessians_iip == true implies func_and_grads_and_hessians!(y, Dy, H, x, p), else y, Dy, H = func_and_grads_and_hessians(x, p).

Alternatively (or additionally), an AbstractAutoDiffBackend can be passed to compute the derivatives if the relevant field isnothing.

Compromise.AlgorithmOptions — Type

AlgorithmOptions(; kwargs...)

Configure the optimization by passing keyword arguments:

log_level::Base.CoreLogging.LogLevel: Control verbosity by setting a min. level for @logmsg. Default: LogLevel(0)

max_iter::Int64: Maximum number of iterations. Default: 500

stop_delta_min::Float64: Stop if the trust region radius is reduced to below stop_delta_min. Default: eps(Float64)

stop_xtol_rel::Float64: Stop if the trial point $xₜ$ is accepted and $‖xₜ - x‖≤ δ‖x‖$. Default: -Inf

stop_xtol_abs::Float64: Stop if the trial point $xₜ$ is accepted and $‖xₜ - x‖≤ ε$. Default: -Inf

stop_ftol_rel::Float64: Stop if the trial point $xₜ$ is accepted and $‖f(xₜ) - f(x)‖≤ δ‖f(x)‖$. Default: -Inf

stop_ftol_abs::Float64: Stop if the trial point $xₜ$ is accepted and $‖f(xₜ) - f(x)‖≤ ε$. Default: -Inf

stop_crit_tol_abs::Float64: Stop if for the approximate criticality it holds that $χ̂(x) <= ε$ and for the feasibility that $θ <= δ$. Default: eps(Float64)

stop_theta_tol_abs::Float64: Stop if for the approximate criticality it holds that $χ̂(x) <= ε$ and for the feasibility that $θ <= δ$. Default: eps(Float64)

stop_max_crit_loops::Int64: Stop after the criticality routine has looped stop_max_crit_loops times. Default: 1

eps_crit::Float64: Lower bound for criticality before entering Criticality Routine. Default: 0.1

eps_theta::Float64: Lower bound for feasibility before entering Criticality Routine. Default: 0.1

crit_B::Float64: At the end of the Criticality Routine the radius is possibly set to crit_B * χ. Default: 1000

crit_M::Float64: Criticality Routine runs until Δ ≤ crit_M * χ. Default: 3000

crit_alpha::Float64: Trust region shrinking factor in criticality loops. Default: 0.5

delta_init::Float64: Initial trust region radius. Default: 0.5

delta_max::Float64: Maximum trust region radius. Default: 2 ^ 5 * delta_init

gamma_shrink_much::Float64: Most severe trust region reduction factor. Default: 0.1

gamma_shrink::Float64: Trust region reduction factor. Default: 0.5

gamma_grow::Float64: Trust region enlargement factor. Default: 2.0

strict_acceptance_test::Bool: Whether to require all objectives to be reduced or not. Default: true

nu_accept::Float64: Acceptance threshold. Default: 0.01

nu_success::Float64: Success threshold. Default: 0.9

c_delta::Float64: Factor for normal step compatibility test. The smaller c_delta, the stricter the test. Default: 0.7

c_mu::Float64: Factor for normal step compatibility test. The smaller c_mu, the stricter the test for small radii. Default: 100.0

mu::Float64: Exponent for normal step compatibility test. The larger mu, the stricter the test for small radii. Default: 0.01

kappa_theta::Float64: Factor in the model decrease condition. Default: 0.0001

psi_theta::Float64: Exponent (for constraint violation) in the model decrease condition. Default: 2.0

scaler_cfg::Symbol: Configuration to determine variable scaling (if model supports it). Either :box or :none. Default: :box

step_config::Any: Configuration object for descent and normal step computation. Default: SteepestDescentConfig()

nl_opt::Symbol: NLopt algorithm symbol for restoration phase. Default: :LN_COBYLA

Compromise.MutableMOP — Type

MutableMOP(; num_vars, kwargs...)

Initialize a multi-objective problem with num_vars variables.

Functions

There can be exactly one (possibly vector-valued) objective function, one nonlinear equality constraint function, and one nonlinear inequality constraint function. For now, they have to be of type NonlinearFunction. You could provide these functions with the keyword-arguments objectives, nl_eq_constraints or nl_ineq_constraints or set the fields of the same name. To conveniently add user-provided functions, there are helper functions, like add_objectives!.

LinearConstraints

Box constraints are defined by the vectors lb and ub. Linear equality constraints $Ex=c$ are defined by the matrix E and the vector c. Inequality constraints read $Ax≤b$ and use A and b.

Surrogate Configuration

Use the keyword-arguments mcfg_objectives to provide an AbstractSurrogateModelConfig to define how the objectives should be modelled. By default, we assume ExactModelConfig(), which requires differentiable objectives.

Compromise.ValueArrays — Type

A struct holding values computed for or derived from an AbstractMOP.

Compromise.add_function! — Function

add_function!(func_field, mop, op, model_cfg; dim_out, backend=NoBackend())

Add the operator op to mop at func_field and use model configuration model_cfg. Keyword argument dim_out::Int is mandatory. E.g., add_function!(:objectives, mop, op, :rbf; dim_out=2) adds op as the bi-valued objective to mop.

Compromise.add_nl_eq_constraints! — Function

add_nl_eq_constraints!(mop::MutableMOP, func, model_cfg=nothing; 
    dim_out::Int, kwargs...)

Set function func to return the nonlinear equality constraints vector of mop. Argument model_cfg is optional and specifies the surrogate models for func. Can be nothing, a Symbol (:exact, :rbf, taylor1, taylor2), or an AbstractSurrogateModelConfig object.

All functions can be in-place, see keyword arguments func_iip.

Keyword argument dim_out is mandatory and corresponds to the length of the result vector. The other kwargs... are passed to the inner AbstractNonlinearOperator as is. For options and defaults see NonlinearParametricFunction.

Compromise.add_nl_eq_constraints! — Function

add_nl_eq_constraints!(mop::MutableMOP, func, grads, func_and_grads, model_cfg=nothing; 
    dim_out::Int, kwargs...)

Set function func to return the nonlinear equality constraints vector of mop. Argument model_cfg is optional and specifies the surrogate models for func. Can be nothing, a Symbol (:exact, :rbf, taylor1, taylor2), or an AbstractSurrogateModelConfig object. grads should be a function mapping a vector to the transposed jacobian of func, while func_and_grads returns a primal vector and the gradients at the same time.

All functions can be in-place, see keyword arguments func_iip, grads_iip and func_and_grads_iip.

Keyword argument dim_out is mandatory and corresponds to the length of the result vector. The other kwargs... are passed to the inner AbstractNonlinearOperator as is. For options and defaults see NonlinearParametricFunction.

Compromise.add_nl_eq_constraints! — Function

add_nl_eq_constraints!(mop::MutableMOP, func, grads, model_cfg=nothing; 
    dim_out::Int, kwargs...)

Set function func to return the nonlinear equality constraints vector of mop. Argument model_cfg is optional and specifies the surrogate models for func. Can be nothing, a Symbol (:exact, :rbf, taylor1, taylor2), or an AbstractSurrogateModelConfig object. grads should be a function mapping a vector to the transposed jacobian of func.

All functions can be in-place, see keyword arguments func_iip and grads_iip.

Keyword argument dim_out is mandatory and corresponds to the length of the result vector. The other kwargs... are passed to the inner AbstractNonlinearOperator as is. For options and defaults see NonlinearParametricFunction.

Compromise.add_nl_ineq_constraints! — Function

add_nl_ineq_constraints!(mop::MutableMOP, func, model_cfg=nothing; 
    dim_out::Int, kwargs...)

Set function func to return the nonlinear inequality constraints vector of mop. Argument model_cfg is optional and specifies the surrogate models for func. Can be nothing, a Symbol (:exact, :rbf, taylor1, taylor2), or an AbstractSurrogateModelConfig object.

All functions can be in-place, see keyword arguments func_iip.

Keyword argument dim_out is mandatory and corresponds to the length of the result vector. The other kwargs... are passed to the inner AbstractNonlinearOperator as is. For options and defaults see NonlinearParametricFunction.

Compromise.add_nl_ineq_constraints! — Function

add_nl_ineq_constraints!(mop::MutableMOP, func, grads, func_and_grads, model_cfg=nothing; 
    dim_out::Int, kwargs...)

Set function func to return the nonlinear inequality constraints vector of mop. Argument model_cfg is optional and specifies the surrogate models for func. Can be nothing, a Symbol (:exact, :rbf, taylor1, taylor2), or an AbstractSurrogateModelConfig object. grads should be a function mapping a vector to the transposed jacobian of func, while func_and_grads returns a primal vector and the gradients at the same time.

All functions can be in-place, see keyword arguments func_iip, grads_iip and func_and_grads_iip.

Keyword argument dim_out is mandatory and corresponds to the length of the result vector. The other kwargs... are passed to the inner AbstractNonlinearOperator as is. For options and defaults see NonlinearParametricFunction.

Compromise.add_nl_ineq_constraints! — Function

add_nl_ineq_constraints!(mop::MutableMOP, func, grads, model_cfg=nothing; 
    dim_out::Int, kwargs...)

Set function func to return the nonlinear inequality constraints vector of mop. Argument model_cfg is optional and specifies the surrogate models for func. Can be nothing, a Symbol (:exact, :rbf, taylor1, taylor2), or an AbstractSurrogateModelConfig object. grads should be a function mapping a vector to the transposed jacobian of func.

All functions can be in-place, see keyword arguments func_iip and grads_iip.

Keyword argument dim_out is mandatory and corresponds to the length of the result vector. The other kwargs... are passed to the inner AbstractNonlinearOperator as is. For options and defaults see NonlinearParametricFunction.

Compromise.add_objectives! — Function

add_objectives!(mop::MutableMOP, func, model_cfg=nothing; 
    dim_out::Int, kwargs...)

Set function func to return the objectives vector of mop. Argument model_cfg is optional and specifies the surrogate models for func. Can be nothing, a Symbol (:exact, :rbf, taylor1, taylor2), or an AbstractSurrogateModelConfig object.

All functions can be in-place, see keyword arguments func_iip.

Keyword argument dim_out is mandatory and corresponds to the length of the result vector. The other kwargs... are passed to the inner AbstractNonlinearOperator as is. For options and defaults see NonlinearParametricFunction.

Compromise.add_objectives! — Function

add_objectives!(mop::MutableMOP, func, grads, func_and_grads, model_cfg=nothing; 
    dim_out::Int, kwargs...)

Set function func to return the objectives vector of mop. Argument model_cfg is optional and specifies the surrogate models for func. Can be nothing, a Symbol (:exact, :rbf, taylor1, taylor2), or an AbstractSurrogateModelConfig object. grads should be a function mapping a vector to the transposed jacobian of func, while func_and_grads returns a primal vector and the gradients at the same time.

All functions can be in-place, see keyword arguments func_iip, grads_iip and func_and_grads_iip.

Keyword argument dim_out is mandatory and corresponds to the length of the result vector. The other kwargs... are passed to the inner AbstractNonlinearOperator as is. For options and defaults see NonlinearParametricFunction.

Compromise.add_objectives! — Function

add_objectives!(mop::MutableMOP, func, grads, model_cfg=nothing; 
    dim_out::Int, kwargs...)

Set function func to return the objectives vector of mop. Argument model_cfg is optional and specifies the surrogate models for func. Can be nothing, a Symbol (:exact, :rbf, taylor1, taylor2), or an AbstractSurrogateModelConfig object. grads should be a function mapping a vector to the transposed jacobian of func.

All functions can be in-place, see keyword arguments func_iip and grads_iip.

Keyword argument dim_out is mandatory and corresponds to the length of the result vector. The other kwargs... are passed to the inner AbstractNonlinearOperator as is. For options and defaults see NonlinearParametricFunction.

Compromise.diff_mod! — Method

Evaluate the model gradients of mod at x and store results in mod_vals::SurrogateValueArrays.

Compromise.eval_and_diff_mod! — Method

Evaluate and differentiate mod at x and store results in mod_vals::SurrogateValueArrays.

Compromise.eval_mod! — Method

Evaluate the models mod at x and store results in mod_vals::SurrogateValueArrays.

Compromise.lin_cons! — Method

lin_cons!(residual_vector, prod_cache, constraint_data, x)

Given a linear constraint A*x .<= b or A*x .== b, compute the product A*x and store the result in prod_cache, and also compute A*x .- b and store the result in residual_vector. constraint_data should either be the tuple (A,b)::Tuple{RMat,RVec} or nothing.

Compromise.scale! — Method

Scale ξ and set x according to x = T*ξ + t.

Compromise.scale_eq! — Method

Make Aξ + b ? 0 applicable in scaled domain via A(inv(T)*x - inv(T)*t) + b ? 0.

Compromise.set_linear_constraints! — Method

set_linear_constraints!(opt, affine_vec, var_vec, mat, rhs, ctype)

Add linear (in-)equality constraints to JuMP model opt.

If ctype is :eq, then the constraints read affine_vec + mat * var_vec .== rhs.

Otherwise, they read affine_vec + mat * var_vec .<= rhs.

This helper function returns a JuMP expression for the matrix-vector-product mat*var_vec.

Compromise.unscale! — Method

Unscale x and set ξ according to ξ = inv(T)*x - inv(T)*t.